Reworked the global structure of the package.
Use our own structure instead of OpenNI context directly.

Use the extend structure to store a few global string constants for
use in callbacks. No need to alloocate a new Tcl_Obj* everytime.

Extended 'user' generator class with methods for setting and unset
callbacks, and invoking it from the C level.

This required extension of the instance structure as well, to keep the
callback command prefixes around, and the OpenNI tokens for the
callbacks.