INCOMPATIBILITY.   Renamed all the events using camelCase to dashed form, i.e camel-case.
  Added a prefix to some them for better distinctibility.

  Reordered the first two arguments from "obj event" to "event
  obj". That allows later to use objects or ensemble as callback
  commands, with the event then used as method.  Much easier dispatch.

Reworked the internals generating the callback C code to ensure that
the dashes do not cause mis-compilation.

Updated all examples to use new argument order and names.

Updated the documentation for the new argument order and names, with
more templating, i.e consistent description, in text, and style.

Regenerated the embedded documentation.