Modified all methods returning images to handle truncation of input data without a panic, returning and incomplete image instead.
This has been seen so far only for depth images.  Time until occurence
is highly variant.  So far the issue has been seen within 5 minutes,
11 minutes, and 30 minutes, all via 'view-skeleton'. Not seen so far
for 'view-depth'.

The moment it occurs probability for reoccurence rapidly climps to
100%. I.e. after first occurence it becomes roughly 50% immediately
100% in short order. Skeleton tracking in NITE seems to get the
truncated images as well, as tracking does not seem to start when
moving in front of the sensor.

It seems that the best way of dealing with the issue is to exit the
process and restart it, cleaning the broken state. If that is done I
would recommend to add timing commands to the restart loop as well, to
gain a larger record of time to occurence from start of process.